Revert "systemd-logind.service: set Type=notify"
This reverts commit a4962513c555fe3ac4b5bebf97a71701361a45b0.

logind.service is a D-Bus service, hence we should use the dbus name as
indication that we are up. Type=dbus is implied if BusName= is
specified, as it is in this case.

This removes a warning that is printed because a BusName= is specified
for a Type=notify unit.